+++ src/Api.php
@@ -111,6 +111,7 @@ class Api
       if ($authenticator !== null) {
         $authenticateArgs = new AuthenticateArgs();
         $authenticateArgs->request = $request;
+        $authenticateArgs->user = $this->wire->user;
         $authenticateArgs->event = $event;
         $authenticator->authenticate($authenticateArgs);
       }
@@ -146,6 +147,7 @@ class Api
       if (!empty($beforeHooks)) {
         $hookReturnBefore = new RequestHookReturnBefore();
         $hookReturnBefore->request = $request;
+        $hookReturnBefore->user = $this->wire->user;
         $hookReturnBefore->event = $event;
         $hookReturnBefore->handler = $handler;
         $hookReturnBefore->endpoint = $result->endpoint;
@@ -164,6 +166,7 @@ class Api
       try {
         $handlerArgs = new EndpointHandlerArgs();
         $handlerArgs->request = $request;
+        $handlerArgs->user = $this->wire->user;
         $handlerArgs->event = $event;
 
         $response = call_user_func($handler, $handlerArgs);
@@ -185,6 +188,7 @@ class Api
           $exceptionHandlerArgs = new ExceptionHandlerArgs();
           $exceptionHandlerArgs->exception = $e;
           $exceptionHandlerArgs->request = $request;
+          $exceptionHandlerArgs->user = $this->wire->user;
           $exceptionHandlerArgs->event = $event;
           $exceptionHandlerArgs->endpoint = $result->endpoint;
           $exceptionHandlerArgs->service = $result->service;
@@ -223,6 +227,7 @@ class Api
       if (!empty($afterHooks)) {
         $hookReturnAfter = new RequestHookReturnAfter();
         $hookReturnAfter->request = $request;
+        $hookReturnAfter->user = $this->wire->user;
         $hookReturnAfter->event = $event;
         $hookReturnAfter->response = $response;
         $hookReturnAfter->endpoint = $result->endpoint;
@@ -247,6 +252,7 @@ class Api
       if (!empty($errorHooks)) {
         $hookReturnError = new ErrorHookReturn();
         $hookReturnError->request = $request;
+        $hookReturnError->user = $this->wire->user;
         $hookReturnError->event = $event;
         $hookReturnError->endpoint = $result->endpoint;
         $hookReturnError->service = $result->service;